Share the ** -capable glob matcher through pkg/facts
The engine has had a matcher that understands ** since the ignore globs
were directory-scoped — vendor/**, **/build/**, **/*_test.go,
prefix/**/fileglob — and its own comment names the bug it exists to
avoid: filepath.Match reads ** as *, which cannot cross a /.

It was unexported and in internal/, so the enterprise performance
analyzer could not reach it and grew a second, weaker implementation on
path.Match. That one silently matched nothing for every ** pattern its
own docs advertised.

Move matchGlob/matchAnyGlob/matchDirScopedGlob to internal/facts, beside
the other shared path predicates (IsTestPath), and re-export MatchGlob /
MatchAnyGlob from pkg/facts. internal/engine keeps thin aliases, so its
call sites and both its existing glob test files are untouched.

Pure code move: no behaviour change, no facts emitted, no cacheVersion
bump. TestGolden passing unchanged is the proof — if a golden moved, the
move altered what the walker ignores.
